refactor: chat-actions, renderMain
