+++ frontend/src/routes/conversation.tsx
@@ -1,7 +1,7 @@
 import { useDisclosure } from "@heroui/react";
 import React from "react";
 import { useNavigate } from "react-router";
-import { useDispatch, useSelector } from "react-redux";
+import { useDispatch } from "react-redux";
 
 import { useConversationId } from "#/hooks/use-conversation-id";
 import { Controls } from "#/components/features/controls/controls";
@@ -9,15 +9,10 @@ import { clearTerminal } from "#/state/command-slice";
 import { useEffectOnce } from "#/hooks/use-effect-once";
 import { clearJupyter } from "#/state/jupyter-slice";
 
-import { ChatInterface } from "../components/features/chat/chat-interface";
 import { WsClientProvider } from "#/context/ws-client-provider";
 import { EventHandler } from "../wrapper/event-handler";
 import { useConversationConfig } from "#/hooks/query/use-conversation-config";
 
-import {
-  Orientation,
-  ResizablePanel,
-} from "#/components/layout/resizable-panel";
 import Security from "#/components/shared/modals/security/security";
 import { useActiveConversation } from "#/hooks/query/use-active-conversation";
 import { useSettings } from "#/hooks/query/use-settings";
@@ -27,10 +22,8 @@ import OpenHands from "#/api/open-hands";
 import { useIsAuthed } from "#/hooks/query/use-is-authed";
 import { ConversationSubscriptionsProvider } from "#/context/conversation-subscriptions-provider";
 import { useUserProviders } from "#/hooks/use-user-providers";
-import { ConversationTabs } from "#/components/features/conversation/conversation-tabs";
 import { ChatActions } from "#/components/features/chat/chat-actions";
-import { RootState } from "#/store";
-import { cn } from "#/utils/utils";
+import { ConversationMain } from "#/components/features/conversation/conversation-main";
 
 function AppContent() {
   useConversationConfig();
@@ -45,12 +38,6 @@ function AppContent() {
 
   // Set the document title to the conversation title when available
   useDocumentTitleFromState();
-
-  const [width, setWidth] = React.useState(window.innerWidth);
-
-  const isRightPanelShown = useSelector(
-    (state: RootState) => state.conversation.isRightPanelShown,
-  );
 
   React.useEffect(() => {
     if (isFetched && !conversation && isAuthed) {
@@ -76,64 +63,11 @@ function AppContent() {
     dispatch(clearJupyter());
   });
 
-  function handleResize() {
-    setWidth(window.innerWidth);
-  }
-
-  React.useEffect(() => {
-    window.addEventListener("resize", handleResize);
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, []);
-
   const {
     isOpen: securityModalIsOpen,
     onOpen: onSecurityModalOpen,
     onOpenChange: onSecurityModalOpenChange,
   } = useDisclosure();
-
-  function renderMain() {
-    if (width <= 1024) {
-      return (
-        <div className="flex flex-col gap-3 overflow-auto w-full">
-          <div
-            className={cn(
-              "rounded-xl overflow-hidden border border-neutral-600 w-full bg-base-secondary min-h-[494px]",
-              !isRightPanelShown && "h-full",
-            )}
-          >
-            <ChatInterface />
-          </div>
-          {isRightPanelShown && (
-            <div className="h-full w-full min-h-[494px]">
-              <ConversationTabs />
-            </div>
-          )}
-        </div>
-      );
-    }
-    if (isRightPanelShown) {
-      return (
-        <ResizablePanel
-          orientation={Orientation.HORIZONTAL}
-          className="grow h-full min-h-0 min-w-0"
-          initialSize={500}
-          firstClassName="rounded-xl overflow-hidden border border-neutral-600 bg-base-secondary"
-          secondClassName="flex flex-col overflow-hidden"
-          firstChild={<ChatInterface />}
-          secondChild={<ConversationTabs />}
-        />
-      );
-    }
-    return (
-      <div className="flex flex-col gap-3 overflow-auto w-full h-full">
-        <div className="rounded-xl overflow-hidden border border-neutral-600 w-full h-full bg-base-secondary">
-          <ChatInterface />
-        </div>
-      </div>
-    );
-  }
 
   return (
     <WsClientProvider conversationId={conversationId}>
@@ -142,7 +76,9 @@ function AppContent() {
           <div data-testid="app-route" className="flex flex-col h-full gap-3">
             <ChatActions />
 
-            <div className="flex h-full overflow-auto">{renderMain()}</div>
+            <div className="flex h-full overflow-auto">
+              <ConversationMain />
+            </div>
 
             <Controls
               setSecurityOpen={onSecurityModalOpen}
